+++ app/build.gradle
@@ -5,6 +5,9 @@ plugins {
     id 'dagger.hilt.android.plugin'
 }
 
+Properties properties = new Properties()
+properties.load(project.rootProject.file("local.properties").newDataInputStream())
+
 android {
     namespace 'com.weit.odya'
     compileSdk 33
@@ -15,6 +18,7 @@ android {
         targetSdk 33
         versionCode 1
         versionName "1.0"
+        buildConfigField "String", "KAKAO_LOGIN_KEY", properties["KAKAO_LOGIN_KEY"]
     }
 
     buildTypes {
@@ -46,6 +50,8 @@ dependencies {
 
     implementation 'com.google.dagger:hilt-android:2.45'
     kapt 'com.google.dagger:hilt-android-compiler:2.45'
+
+    implementation "com.kakao.sdk:v2-user:2.14.0"
 
     implementation 'com.orhanobut:logger:2.2.0'
 }
