+++ app/src/main/java/com/weit/odya/HiltApplication.kt
@@ -2,22 +2,18 @@ package com.weit.odya
 
 import android.app.Application
 import androidx.appcompat.app.AppCompatDelegate
+import com.kakao.sdk.common.KakaoSdk
 import com.orhanobut.logger.AndroidLogAdapter
 import com.orhanobut.logger.Logger
-import com.weit.data.util.ContextProvider
 import dagger.hilt.android.HiltAndroidApp
-import javax.inject.Inject
 
 @HiltAndroidApp
 class HiltApplication : Application() {
-
-    @Inject
-    lateinit var contextProvider: ContextProvider
 
     override fun onCreate() {
         super.onCreate()
         Logger.addLogAdapter(AndroidLogAdapter())
         AppCompatDelegate.setDefaultNightMode(AppCompatDelegate.MODE_NIGHT_YES)
-        contextProvider.start(this)
+        KakaoSdk.init(this, BuildConfig.KAKAO_LOGIN_KEY)
     }
 }
