+++ data/src/main/java/com/weit/data/di/AuthModule.kt
@@ -2,7 +2,6 @@ package com.weit.data.di
 
 import android.content.Context
 import com.weit.data.repository.auth.AuthRepositoryImpl
-import com.weit.data.util.ContextProvider
 import com.weit.domain.repository.auth.AuthRepository
 import dagger.Module
 import dagger.Provides
@@ -19,10 +18,5 @@ object AuthModule {
     @Provides
     fun providesAuthRepository(
         @ApplicationContext context: Context,
-        activityProvider: ContextProvider,
-    ): AuthRepository = AuthRepositoryImpl(context, activityProvider)
-
-    @Singleton
-    @Provides
-    fun providesContextProvider(): ContextProvider = ContextProvider()
+    ): AuthRepository = AuthRepositoryImpl(context)
 }
