+++ domain/src/main/java/com/weit/domain/repository/auth/AuthRepository.kt
@@ -1,6 +1,5 @@
 package com.weit.domain.repository.auth
 
 interface AuthRepository {
-    suspend fun loginWithKakao(): Result<Unit>
     suspend fun logout(): Result<Unit>
 }
