+++ domain/src/main/java/com/weit/domain/repository/auth/LoginRepository.kt
@@ -1,0 +1,5 @@
+package com.weit.domain.repository.auth
+
+interface LoginRepository {
+    suspend fun loginWithKakao(): Result<Unit>
+}
