+++ domain/src/main/java/com/weit/domain/usecase/auth/LoginWithKakaoUseCase.kt
@@ -1,10 +1,13 @@
 package com.weit.domain.usecase.auth
 
-import com.weit.domain.repository.auth.AuthRepository
+import com.weit.domain.repository.auth.LoginRepository
 import javax.inject.Inject
 
+/**
+ * 카카오 로그인은 반드시 Activity, Fragment에서 주입
+ */
 class LoginWithKakaoUseCase @Inject constructor(
-    private val repository: AuthRepository,
+    private val repository: LoginRepository,
 ) {
     suspend operator fun invoke(): Result<Unit> = repository.loginWithKakao()
 }
