+++ presentation/src/main/java/com/weit/presentation/ui/login/LoginActivity.kt
@@ -3,23 +3,28 @@ package com.weit.presentation.ui.login
 import android.content.Intent
 import androidx.activity.viewModels
 import com.orhanobut.logger.Logger
+import com.weit.domain.usecase.auth.LoginWithKakaoUseCase
 import com.weit.presentation.databinding.ActivityLoginBinding
 import com.weit.presentation.ui.MainActivity
 import com.weit.presentation.ui.base.BaseActivity
 import com.weit.presentation.ui.util.repeatOnStarted
 import dagger.hilt.android.AndroidEntryPoint
 import kotlinx.coroutines.flow.collectLatest
+import javax.inject.Inject
 
 @AndroidEntryPoint
 class LoginActivity : BaseActivity<ActivityLoginBinding>(
     ActivityLoginBinding::inflate,
 ) {
 
+    @Inject
+    lateinit var loginWithKakaoUseCase: LoginWithKakaoUseCase
+
     private val viewModel: LoginViewModel by viewModels()
 
     override fun initListener() {
         binding.btnLoginKakao.setOnClickListener {
-            viewModel.onLoginWithKakao()
+            viewModel.onLoginWithKakao(loginWithKakaoUseCase)
         }
     }
